Add functionality to create todo
